Rename raidal_observer to radial_painter
