+++ lib/src/charts/widgets/pie_chart.dart
@@ -6,7 +6,7 @@ import 'package:chart_it/src/charts/data/pie/pie_series.dart';
 import 'package:chart_it/src/charts/painters/radial/pie_painter.dart';
 import 'package:chart_it/src/charts/painters/radial/radial_painter.dart';
 import 'package:chart_it/src/charts/widgets/core/radial_charts.dart';
-import 'package:chart_it/src/controllers/radial_observer.dart';
+import 'package:chart_it/src/controllers/radial_controller.dart';
 import 'package:flutter/material.dart';
 
 /// Draws a PieChart or Donut Chart for the Provided Data
